Update to use BigInteger/BigDecimal Submitted by: villard@us.ibm.com
git-svn-id: https://svn.eu.apache.org/repos/asf/xalan/java/branches/xslt20@337279 13f79535-47bb-0310-9956-ffa450edef68

+++ xpath_rwapi/src/org/apache/xpath/rwapi/expression/Literal.java
@@ -55,6 +55,9 @@
  */
 package org.apache.xpath.rwapi.expression;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
+
 import org.apache.xpath.rwapi.XPathException;
 
 /**
@@ -101,12 +104,12 @@ public interface Literal extends Expr {
     /**
      * @throws XPathException when the literal isn't an integer
      */
-    int getIntegerLiteral() throws XPathException;
+    BigInteger getIntegerLiteral() throws XPathException;
    
     /**
      * @throws XPathException when the literal isn't a decimal
      */
-    float getDecimalLiteral() throws XPathException;
+    BigDecimal getDecimalLiteral() throws XPathException;
    
     /**
      * @throws XPathException when the literal isn't a double
